+++ src/ios/ImageEditorController+Flip.swift
@@ -1,8 +1,0 @@
-import UIKit
-
-extension ImageEditorController {
-    
-    func flipHorizontaly(_ image: UIImage) -> UIImage {
-        return image.withHorizontallyFlippedOrientation()
-    }
-}