+++ www/Camera.js
@@ -134,23 +134,23 @@ for (var key in Camera) {
 cameraExport.getPicture = function (successCallback, errorCallback, options) {
     argscheck.checkArgs('fFO', 'Camera.getPicture', arguments);
     options = options || {};
-    var getValue = argscheck.getValue;
+    let getValue = argscheck.getValue;
 
-    var quality = getValue(options.quality, 50);
-    var destinationType = getValue(options.destinationType, Camera.DestinationType.FILE_URI);
-    var sourceType = getValue(options.sourceType, Camera.PictureSourceType.CAMERA);
-    var targetWidth = getValue(options.targetWidth, -1);
-    var targetHeight = getValue(options.targetHeight, -1);
-    var encodingType = getValue(options.encodingType, Camera.EncodingType.JPEG);
-    var mediaType = getValue(options.mediaType, Camera.MediaType.PICTURE);
-    var allowEdit = !!options.allowEdit;
-    var correctOrientation = !!options.correctOrientation;
-    var saveToPhotoAlbum = !!options.saveToPhotoAlbum;
-    var popoverOptions = getValue(options.popoverOptions, null);
-    var cameraDirection = getValue(options.cameraDirection, Camera.Direction.BACK);
+    let quality = getValue(options.quality, 50);
+    let sourceType = getValue(options.sourceType, Camera.PictureSourceType.CAMERA);
+    let destinationType = getValue(options.destinationType, Camera.DestinationType.FILE_URI);
+    let targetWidth = getValue(options.targetWidth, -1);
+    let targetHeight = getValue(options.targetHeight, -1);
+    let encodingType = getValue(options.encodingType, Camera.EncodingType.JPEG);
+    let mediaType = getValue(options.mediaType, Camera.MediaType.PICTURE);
+    let allowEdit = !!options.allowEdit;
+    let correctOrientation = !!options.correctOrientation;
+    let saveToPhotoAlbum = !!options.saveToPhotoAlbum;
+    let cameraDirection = getValue(options.cameraDirection, Camera.Direction.BACK);
 
-    var args = [quality, targetWidth, targetHeight, encodingType, allowEdit, correctOrientation, 
-        saveToPhotoAlbum, cameraDirection, destinationType, sourceType, mediaType, popoverOptions];
+
+    let args = [{quality, targetWidth, targetHeight, encodingType, allowEdit, correctOrientation, 
+        saveToPhotoAlbum, cameraDirection, destinationType, sourceType, mediaType}];
 
     exec(successCallback, errorCallback, 'Camera', 'takePicture', args);
     // XXX: commented out
@@ -182,11 +182,7 @@ cameraExport.cleanup = function (successCallback, errorCallback) {
     exec(successCallback, errorCallback, 'Camera', 'cleanup', []);
 };
 
-/**
- * TODO: Description of functionality
- */
- cameraExport.editPicture = function (successCallback, errorCallback, options) {
-
+cameraExport.editPicture = function (successCallback, errorCallback, options) {
     argscheck.checkArgs('fFO', 'Camera.editPicture', arguments);
     options = options || {};
     var getValue = argscheck.getValue;
@@ -198,6 +194,16 @@ cameraExport.cleanup = function (successCallback, errorCallback) {
 
     var args = [imageByteArray];
     exec(successCallback, errorCallback, 'Camera', 'editPicture', args);
+};
+
+cameraExport.editURIPicture = function (successCallback, errorCallback, options) {
+    argscheck.checkArgs('fFO', 'Camera.editURIPicture', arguments);
+    options = options || {};
+    
+    let {uri, saveToGallery, includeMetadata} = options;
+    let args = [{uri, saveToGallery, includeMetadata}];
+
+    exec(successCallback, errorCallback, 'Camera', 'editURIPicture', args);
 };
 
 cameraExport.recordVideo = function(successCallback, errorCallback, options) {
@@ -219,8 +225,9 @@ cameraExport.chooseFromGallery = function(successCallback, errorCallback, option
     let mediaType = options.mediaType;
     let allowMultipleSelection = !!options.allowMultipleSelection;
     let includeMetadata = !!options.includeMetadata;
+    let allowEdit = !!options.allowEdit;
 
-    let args = [{mediaType, allowMultipleSelection, includeMetadata}];
+    let args = [{mediaType, allowMultipleSelection, includeMetadata, allowEdit}];
 
     exec(successCallback, errorCallback, 'Camera', 'chooseFromGallery', args);
 }
@@ -235,5 +242,33 @@ cameraExport.playVideo = function(successCallback, errorCallback, options){
 
     exec(successCallback, errorCallback, 'Camera', 'playVideo', args);
 }
+
+cameraExport.takePicture = function (successCallback, errorCallback, options) {
+    argscheck.checkArgs('fFO', 'Camera.takePicture', arguments);
+    options = options || {};
+    let getValue = argscheck.getValue;
+
+    let quality = getValue(options.quality, 50);
+    let sourceType = getValue(options.sourceType, Camera.PictureSourceType.CAMERA);
+    let destinationType = getValue(options.destinationType, Camera.DestinationType.FILE_URI);
+    let targetWidth = getValue(options.targetWidth, -1);
+    let targetHeight = getValue(options.targetHeight, -1);
+    let encodingType = getValue(options.encodingType, Camera.EncodingType.JPEG);
+    let mediaType = getValue(options.mediaType, Camera.MediaType.PICTURE);
+    let allowEdit = !!options.allowEdit;
+    let correctOrientation = !!options.correctOrientation;
+    let saveToPhotoAlbum = !!options.saveToPhotoAlbum;
+    let cameraDirection = getValue(options.cameraDirection, Camera.Direction.BACK);
+    let includeMetadata = !!options.includeMetadata;
+    let latestVersion = !!options.latestVersion;
+
+    let args = [{quality, targetWidth, targetHeight, encodingType, allowEdit, correctOrientation, 
+        saveToPhotoAlbum, cameraDirection, destinationType, sourceType, mediaType, includeMetadata,
+        latestVersion}];
+
+    exec(successCallback, errorCallback, 'Camera', 'takePicture', args);
+    // XXX: commented out
+    // return new CameraPopoverHandle();
+};
 
 module.exports = cameraExport;